+++ packages/superset-ui-style/package.json
@@ -32,5 +32,9 @@
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
     "emotion-theming": "^10.0.27"
+  },
+  "devDependencies": {
+    "react": "^16.13.1",
+    "enzyme": "^3.11.0"
   }
 }
